CancelService+timer

+++ app/src/main/java/com/flexship/flexshipcookingass/models/Stages.kt
@@ -5,8 +5,8 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "stages_table")
 data class Stages(
-    @PrimaryKey (autoGenerate = true) var id: Int = 0,
-    val name: String ="",
-    val time: Int =0,
-    val dishId:Int =0
+    @PrimaryKey(autoGenerate = true) var id: Int = 0,
+    val name: String = "",
+    val time: Long = 0L,
+    val dishId: Int = 0
 )
